fixed some styling imports   -- renamed todo-app.js to index.js so that main styling would be auto imported, implemented precustomizations and customizations for bootstrap-loader to allow for a custom build
