+++ app/components/entry/entry-view.html
@@ -1,6 +1,14 @@
 <section class="Entry container-fluid">
+  
   <div class="Entry-wrapper row">
-    <button type="button" name="button"></button>
+    
+    
+    <button 
+    type="button" 
+    name="button"
+    data-ng-click=""
+    data-ng-bind=""
+    ></button>
     <button 
     type="button" 
     name="button"
@@ -10,6 +18,7 @@
     
     <form 
     class="col-xs-12 col-md-3 col-md-offset-3" 
+    data-ng-show=""
     novalidate>
       <div class="form-group">
         <label for=""></label>
