--- app/components/entry/entry.scss
@@ -1,3 +1,0 @@
-body {
-  color: white;
-}
