+++ app/entry.js
@@ -1,12 +1,11 @@
 require('!!file?name=[name].[ext]!./main/index.html');
-require('./style/base.scss');
 
 require('./services');
 
-require('./components/entry/index.js');
+require('./components/entry');
 require('./components/navbar');
 require('./components/item');
 require('./components/list');
 require('./components/board');
 
-require('./main/todo-app');
+require('./main');
