+++ app/main/index.html
@@ -6,7 +6,6 @@
     <link rel="stylesheet" href="bundle.css" charset="utf-8">
   </head>
   <body>
-    {{ 1 + 1 }}
     
     <main data-ng-view></main>
