+++ app/main/main.scss
@@ -1,0 +1,3 @@
+html {
+  position: relative;
+}
